не вызываем solrj с пустым списком

+++ src/ru/org/linux/spring/SearchQueueListener.java
@@ -86,7 +86,9 @@ public class SearchQueueListener {
           }
         });
 
-        solrServer.deleteById(msgids);
+        if (!msgids.isEmpty()) {
+          solrServer.deleteById(msgids);
+        }
       }
     }
   }
